requirements.txt
basicsr==1.4.2
lpips==0.1.4
numpy==2.4.0
Pillow==12.0.0
piq==0.8.0
skimage==0.0
torch==2.1.0+cu118
torchvision==0.16.0+cu118
